# MobyPark/OpenAPI/openapi.yaml
openapi: 3.0.4
info: {}
paths:
  /api/Business:
    post:
      tags:
      - Business
      summary: Creates a new business.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateBusinessDto'
      responses:
        '400':
          description: Invalid IBAN provided
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Business created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadBusinessDto'
        '409':
          description: Address already in use
  /api/Business/{id}:
    patch:
      tags:
      - Business
      summary: Updates an existing business.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchBusinessDto'
          text/json:
            schema:
              $ref: '#/components/schemas/PatchBusinessDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/PatchBusinessDto'
      responses:
        '400':
          description: Invalid IBAN provided
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Business updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadBusinessDto'
        '404':
          description: Business not found
        '409':
          description: New address already taken
    delete:
      tags:
      - Business
      summary: Deletes a business by its ID.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Business deleted
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Business not found
    get:
      tags:
      - Business
      summary: Retrieves a business by its ID.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Found business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadBusinessDto'
        '404':
          description: Not found
  /api/Business/all:
    get:
      tags:
      - Business
      summary: Retrieves a list of all registered businesses.
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: List retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadBusinessDto'
  /api/Business/{address}:
    get:
      tags:
      - Business
      summary: Retrieves a business by its address.
      parameters:
      - name: address
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Found business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadBusinessDto'
        '404':
          description: Not found
...
  /api/BusinessParking/business/{id}:
    get:
      tags:
      - BusinessParkingRegistrations
      summary: Retrieves all registrations linked to a specific business ID.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: List retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadBusinessRegDto'
        '404':
          description: No registrations found for this business
...
  /api/Hotels:
    post:
      tags:
      - Hotels
      summary: Creates a new hotel.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHotelDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateHotelDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateHotelDto'
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Hotel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadHotelDto'
        '404':
          description: Parking lot not found
        '409':
          description: Address or Parking Lot already taken
    patch:
      tags:
      - Hotels
      summary: Updates an existing hotel.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchHotelDto'
          text/json:
            schema:
              $ref: '#/components/schemas/PatchHotelDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/PatchHotelDto'
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Hotel updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatchHotelDto'
        '404':
          description: Hotel not found
        '409':
          description: New address or parking lot already taken
    get:
      tags:
      - Hotels
      summary: Retrieves a list of all hotels.
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: List retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadHotelDto'
  /api/Hotels/{id}:
    delete:
      tags:
      - Hotels
      summary: Deletes a hotel by ID.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Hotel not found
    get:
      tags:
      - Hotels
      summary: Retrieves a hotel by ID.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Hotel found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadHotelDto'
        '404':
          description: Hotel not found
...
components:
  schemas:
...
    CreateBusinessDto:
      required:
      - address
      - iban
      - name
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 1
          type: string
          description: The official registered name of the business
        address:
          minLength: 1
          type: string
          description: The physical visiting address
        iban:
          minLength: 1
          type: string
          description: The bank account number (IBAN format)
      additionalProperties: false
      description: Data Transfer Object for creating a new business
...
    CreateHotelDto:
      required:
      - address
      - hotelParkingLotId
      - iban
      - name
      type: object
      properties:
        name:
          minLength: 1
          type: string
          description: The name of the hotel
        address:
          minLength: 1
          type: string
          description: Physical address
        iban:
          minLength: 1
          type: string
          description: Bank account number (IBAN)
        hotelParkingLotId:
          type: integer
          description: The ID of the parking lot owned by this hotel
          format: int64
      additionalProperties: false
      description: Data for creating a new Hotel.
...
    PatchBusinessDto:
      type: object
      properties:
        name:
          type: string
          description: The official registered name of the business
          nullable: true
        address:
          type: string
          description: The physical visiting address
          nullable: true
        iban:
          type: string
          description: The bank account number (IBAN format)
          nullable: true
      additionalProperties: false
      description: Data Transfer Object for updating existing business information
...
    PatchHotelDto:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the hotel
          format: int64
        name:
          type: string
          description: The new name (optional)
          nullable: true
        address:
          type: string
          description: The new address (optional)
          nullable: true
        iban:
          type: string
          description: The new IBAN (optional)
          nullable: true
        hotelParkingLotId:
          type: integer
          description: The new parking lot ID (optional)
          format: int64
          nullable: true
      additionalProperties: false
      description: Data for updating a Hotel.
...
    ReadBusinessDto:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the business
          format: int64
        name:
          type: string
          description: The official registered name of the business
          nullable: true
        address:
          type: string
          description: The physical visiting address
          nullable: true
        iban:
          type: string
          description: The bank account number (IBAN format)
          nullable: true
      additionalProperties: false
      description: Data Transfer Object for reading business information
    ReadBusinessRegDto:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the registration
          format: int64
        businessId:
          type: integer
          description: The ID of the associated business
          format: int64
        licensePlateNumber:
          type: string
          description: The license plate number
          nullable: true
        active:
          type: boolean
          description: Whether the registration is currently valid
        lastSinceActive:
          type: string
          description: The timestamp when the active status was last changed
          format: date-time
      additionalProperties: false
      description: Details of a business parking registration.
    ReadHotelDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        hotelParkingLotId:
          type: integer
          format: int64
      additionalProperties: false
      description: Details of a Hotel.
...
  securitySchemes:
    Bearer:
      type: http
      description: 'Please enter a valid token (e.g., ''Bearer eyJ...'').

        You can obtain a token by logging in or registering a new account via the
        endpoints.'
      scheme: Bearer
      bearerFormat: JWT
security:
- Bearer: []